+++ doc/whats_new.rst
@@ -129,6 +129,9 @@ API changes summary
    - Sparse matrix support in :class:`sklearn.decomposition.RandomizedPCA`
      is now deprecated in favor of the new ``TruncatedSVD``.
 
+   - :class:`cross_valiation.KFold` and
+     :class:`cross_valiation.StratifiedKFold` now enforce `n_folds >= 2`
+     otherwise a ``ValueError`` is raised. By `Olivier Grisel`_.
 
 .. _changes_0_13_1: